# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sistema de carteiras digitais multi-tenant desenvolvido em Go",
        "title": "Suno Wallets API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte API",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/b3/client/last-sync": {
            "get": {
                "description": "Retorna os timestamps das últimas sincronizações de transações e posições para um CPF específico, útil para controle de incrementais.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B3 Sync"
                ],
                "summary": "Obter últimas sincronizações do cliente",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "12345678901",
                        "description": "CPF do cliente (11 dígitos, apenas números)",
                        "name": "cpf",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Timestamps das últimas sincronizações",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "CPF inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/b3/client/status": {
            "get": {
                "description": "Retorna status detalhado da sincronização de um cliente, incluindo flags de ativação, necessidade de reprocessamento, últimas sincronizações, resultados e contadores de falha.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B3 Sync"
                ],
                "summary": "Obter status de sincronização do cliente",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "12345678901",
                        "description": "CPF do cliente (11 dígitos, apenas números)",
                        "name": "cpf",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status completo de sincronização",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "CPF inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/b3/client/sync-window": {
            "get": {
                "description": "Calcula e retorna informações sobre a janela de sincronização incremental para um cliente, incluindo período de cobertura e estimativas de páginas a processar.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B3 Utilities"
                ],
                "summary": "Janela de sincronização incremental",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "12345678901",
                        "description": "CPF do cliente (11 dígitos, apenas números)",
                        "name": "cpf",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "transactions",
                        "description": "Tipo de dados (transactions ou positions)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-01-01",
                        "description": "Data inicial no formato YYYY-MM-DD",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2024-12-31",
                        "description": "Data final no formato YYYY-MM-DD",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações da janela de sincronização",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parâmetros inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/b3/sync/run": {
            "post": {
                "description": "Executa sincronização incremental com a B3 para atualizar dados de transações e posições. Suporta escopo single (CPF específico) ou tenant (todos os CPFs). Inclui controles de força, dry-run e limite.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B3 Sync"
                ],
                "summary": "Executar sincronização incremental B3",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Parâmetros da sincronização",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/src_api_controllers_b3.runRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado da sincronização com contadores",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parâmetros inválidos ou erro na sincronização",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/b3/test/connection": {
            "get": {
                "description": "Testa a conectividade com a API da B3 usando credenciais do cliente e retorna relatório detalhado de status. Este endpoint é essencial para validar se as credenciais estão funcionando antes de executar operações de sincronização.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B3 Test"
                ],
                "summary": "Teste de conexão com a B3",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "12345678901",
                        "description": "CPF do cliente (11 dígitos, apenas números)",
                        "name": "cpf",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-01",
                        "description": "Período para teste no formato YYYY-MM",
                        "name": "yearMonth",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conexão bem-sucedida",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parâmetros inválidos (CPF ou yearMonth)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Falha na conexão com B3",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "src_api_controllers_b3.runRequest": {
            "type": "object",
            "properties": {
                "assetTypes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "equity"
                    ]
                },
                "cpf": {
                    "type": "string",
                    "example": "12345678901"
                },
                "dataTypes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "transactions",
                        "positions"
                    ]
                },
                "dryRun": {
                    "type": "boolean"
                },
                "force": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer",
                    "example": 100
                },
                "scope": {
                    "type": "string",
                    "example": "single|tenant"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "TenantID": {
            "description": "ID do inquilino para multi-tenancy",
            "type": "apiKey",
            "name": "X-Tenant-ID",
            "in": "header"
        }
    }
}